+++ R/get_data.R
@@ -16,18 +16,23 @@
 #'   x <- get_data("regioner")
 #'   head(x)
 #' }
-get_data <- function(section,
-                     as_list = FALSE, ...) {
+get_data <- function(section, as_list = FALSE, ...) {
+  warn_api_deprecation()
+
   if (!section %in% available_sections()) {
-    cli::cli_abort("You have provided section {.var {section}}.
-                   It is not available through the api")
+    cli::cli_abort(
+      "You have provided section {.var {section}}.
+                   It is not available through the api"
+    )
   }
 
   params <- rlang::list2(...)
 
   if (!connection_check()) {
-    cli::cli_alert_warning("You do not have access to api.dataforsyningen.dk.
-        Please check your connection settings.")
+    cli::cli_alert_warning(
+      "You do not have access to api.dataforsyningen.dk.
+        Please check your connection settings."
+    )
     return(NULL) # Exit early if no connection is detected
   }